full working prototype of data acquisition into db from cobot api task

+++ manage.py
@@ -1,14 +1,26 @@
 #!/usr/bin/env python
+# -*- coding: utf-8 -*-
+from __future__ import print_function
+import sys
 import unittest
-from flask.ext.script import Shell, Server, Manager
-from app import app, db, models
-from app.tasks import get_data_of_a_day
 import traceback
+from flask.ext.migrate import Migrate, MigrateCommand
+from flask.ext.script import Shell, Server, Manager, prompt_bool
+from flask.ext.script.commands import InvalidCommand
+from app import app, db, models
+from app.tasks import (start_data_task_of_day,
+                       start_data_task_of_duration,
+                       start_report_task_of_month,
+                       start_report_task_of_duration)
 
 manager = Manager(app)
 
-# add runserver command to serve this app
+# add `runserver` command to serve this app
 manager.add_command("runserver", Server())
+
+# add `db` command to handle database migrations
+migrate = Migrate(app, db)
+manager.add_command('db', MigrateCommand)
 
 
 # add shell command to open python shell
@@ -21,26 +33,55 @@ manager.add_command("shell", Shell(make_context=_make_context))
 @manager.command
 def create_db():
     """Creates database tables from sqlalchemy models"""
-    db.create_all()
-    print 'Database tables successfully created.'
+    if prompt_bool("Are you sure you want to create new tables"):
+        db.create_all()
+        print('Database tables successfully created.')
 
 
 @manager.command
 def drop_db():
     """Drops database tables"""
-    db.drop_all()
-    print 'Database tables successfully droped.'
+    if prompt_bool("Are you sure you want to lose all your data"):
+        db.drop_all()
+        print('Database tables successfully dropped.')
 
 
-@manager.option('-d', '--date', dest='date', default=None)
-def run_task(date):
+@manager.option('-sd', '--startDate', dest='start_date', default=None,
+                help="start date of crawl in 'YYYY-MM-DD' format")
+@manager.option('-ed', '--endDate', dest='end_date', default=None,
+                help="end date of crawl in 'YYYY-MM-DD' format")
+def run_task_data(start_date, end_date):
     """Runs a task to get and insert data from cobot api"""
     try:
-        if date:
-            get_data_of_a_day(date)
-            print '===> Task Completed\n'
+        if start_date and end_date:
+            start_data_task_of_duration(start_date, end_date)
+        elif start_date:
+            start_data_task_of_day(start_date)
         else:
-            print 'date argument is not provided'
+            print('Check argument options, type command with --help')
+            return
+
+        print('===> Task Completed')
+    except Exception:
+        traceback.print_exc()
+
+
+@manager.option('-sd', '--startDate', dest='start_date', default=None,
+                help="start date of crawl in 'YYYY-MM' format")
+@manager.option('-ed', '--endDate', dest='end_date', default=None,
+                help="end date of crawl in 'YYYY-MM' format")
+def run_task_report(start_date, end_date):
+    """Runs a task to calculate member report metrices from database data"""
+    try:
+        if start_date and end_date:
+            start_report_task_of_duration(start_date, end_date)
+        elif start_date:
+            start_report_task_of_month(start_date)
+        else:
+            print('Check argument options, type command with --help')
+            return
+
+        print('===> Task Completed')
     except Exception:
         traceback.print_exc()
 
@@ -57,4 +98,8 @@ def test():
 
 
 if __name__ == '__main__':
-    manager.run()
+    try:
+        manager.run()
+    except InvalidCommand as err:
+        print(err, file=sys.stderr)
+        sys.exit(1)
